+++ cuda/src/main.cu
@@ -16,12 +16,12 @@ void move_neccessary_data_before_printing(Simulation *sim) {
     CHECK(cudaMemcpy(&sim->totalCollisions, sim->d_totalCollisions, sizeof(unsigned long long), cudaMemcpyDeviceToHost))
 //    CHECK(cudaMemcpy(sim->P, sim->d_P, sim->NP * 6 * sizeof(double), cudaMemcpyDeviceToHost));
 //
-    CHECK(cudaMemcpy(sim->P.x, sim->d_P.x, sim->NP * sizeof(double), cudaMemcpyDeviceToHost));
-    CHECK(cudaMemcpy(sim->P.y, sim->d_P.y, sim->NP * sizeof(double), cudaMemcpyDeviceToHost));
-    CHECK(cudaMemcpy(sim->P.z, sim->d_P.z, sim->NP * sizeof(double), cudaMemcpyDeviceToHost));
-    CHECK(cudaMemcpy(sim->P.vx, sim->d_P.vx, sim->NP * sizeof(double), cudaMemcpyDeviceToHost));
-    CHECK(cudaMemcpy(sim->P.vy, sim->d_P.vy, sim->NP * sizeof(double), cudaMemcpyDeviceToHost));
-    CHECK(cudaMemcpy(sim->P.vz, sim->d_P.vz, sim->NP * sizeof(double), cudaMemcpyDeviceToHost));
+    CHECK(cudaMemcpy(sim->P.x, sim->d_P.x, sim->NP * sizeof(float), cudaMemcpyDeviceToHost));
+    CHECK(cudaMemcpy(sim->P.y, sim->d_P.y, sim->NP * sizeof(float), cudaMemcpyDeviceToHost));
+    CHECK(cudaMemcpy(sim->P.z, sim->d_P.z, sim->NP * sizeof(float), cudaMemcpyDeviceToHost));
+    CHECK(cudaMemcpy(sim->P.vx, sim->d_P.vx, sim->NP * sizeof(float), cudaMemcpyDeviceToHost));
+    CHECK(cudaMemcpy(sim->P.vy, sim->d_P.vy, sim->NP * sizeof(float), cudaMemcpyDeviceToHost));
+    CHECK(cudaMemcpy(sim->P.vz, sim->d_P.vz, sim->NP * sizeof(float), cudaMemcpyDeviceToHost));
 
     CHECK(cudaMemcpy(sim->samples, sim->d_samples, SAMPLES_SZ, cudaMemcpyDeviceToHost));
 }
